+++ src/polyfills.js
@@ -19,7 +19,7 @@
     }
 }(this, function () {
 
-/** PrivateFunction: Function.prototype.bind
+/** Function: Function.prototype.bind
  *  Bind a function to an instance.
  *
  *  This Function object extension method creates a bound method similar
@@ -46,16 +46,13 @@ if (!Function.prototype.bind) {
         var _slice = Array.prototype.slice;
         var _concat = Array.prototype.concat;
         var _args = _slice.call(arguments, 1);
-
         return function () {
-            return func.apply(obj ? obj : this,
-                              _concat.call(_args,
-                                           _slice.call(arguments, 0)));
+            return func.apply(obj ? obj : this, _concat.call(_args, _slice.call(arguments, 0)));
         };
     };
 }
 
-/** PrivateFunction: Array.isArray
+/** Function: Array.isArray
  *  This is a polyfill for the ES5 Array.isArray method.
  */
 if (!Array.isArray) {
@@ -64,7 +61,7 @@ if (!Array.isArray) {
     };
 }
 
-/** PrivateFunction: Array.prototype.indexOf
+/** Function: Array.prototype.indexOf
  *  Return the index of an object in an array.
  *
  *  This function is not supplied by some JavaScript implementations, so
@@ -78,12 +75,9 @@ if (!Array.isArray) {
  *  Returns:
  *    The index of elt in the array or -1 if not found.
  */
-if (!Array.prototype.indexOf)
-    {
-        Array.prototype.indexOf = function(elt /*, from*/)
-        {
+if (!Array.prototype.indexOf) {
+        Array.prototype.indexOf = function(elt /*, from*/) {
             var len = this.length;
-
             var from = Number(arguments[1]) || 0;
             from = (from < 0) ? Math.ceil(from) : Math.floor(from);
             if (from < 0) {
@@ -95,8 +89,63 @@ if (!Array.prototype.indexOf)
                     return from;
                 }
             }
-
             return -1;
         };
     }
 }));
+
+
+/** Function: Array.prototype.forEach
+ *  
+ *  This function is not available in IE < 9
+ *  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/forEach
+ */
+if (!Array.prototype.forEach) {
+    Array.prototype.forEach = function(callback, thisArg) {
+        var T, k;
+        if (this === null) {
+            throw new TypeError(' this is null or not defined');
+        }
+
+        // 1. Let O be the result of calling toObject() passing the
+        // |this| value as the argument.
+        var O = Object(this);
+        // 2. Let lenValue be the result of calling the Get() internal
+        // method of O with the argument "length".
+        // 3. Let len be toUint32(lenValue).
+        var len = O.length >>> 0;
+        // 4. If isCallable(callback) is false, throw a TypeError exception. 
+        // See: http://es5.github.com/#x9.11
+        if (typeof callback !== "function") {
+            throw new TypeError(callback + ' is not a function');
+        }
+        // 5. If thisArg was supplied, let T be thisArg; else let
+        // T be undefined.
+        if (arguments.length > 1) {
+            T = thisArg;
+        }
+        // 6. Let k be 0
+        k = 0;
+        // 7. Repeat, while k < len
+        while (k < len) {
+            var kValue;
+            // a. Let Pk be ToString(k).
+            //        This is implicit for LHS operands of the in operator
+            // b. Let kPresent be the result of calling the HasProperty
+            //        internal method of O with argument Pk.
+            //        This step can be combined with c
+            // c. If kPresent is true, then
+            if (k in O) {
+                // i. Let kValue be the result of calling the Get internal
+                // method of O with argument Pk.
+                kValue = O[k];
+                // ii. Call the Call internal method of callback with T as
+                // the this value and argument list containing kValue, k, and O.
+                callback.call(T, kValue, k, O);
+            }
+            // d. Increase k by 1.
+            k++;
+        }
+        // 8. return undefined
+    };
+}